requirements.txt
pandas>=2.2.0
numpy>=1.23.0
pydantic>=1.10.0
yfinance>=0.2.18
twelvedata>=1.2.25
requests>=2.31.0
openpyxl>=3.1.2
tqdm>=4.66.0
python-dotenv>=1.0.0
